==> 01_visualize_tracks.Rmd <==
---
title: "Prepae metadata file for asthma data"
author: "TB"
date: "7/28/2017"
output: html_document
---
```{r}
library("tidyverse")
library("stringr")
library("readr")
library("plotly")
library("ggplot2")
library("neutrominer")
library("magrittr")
```

# Load data
Define analysis folder
```{r}
analysis_folder <- "../../analysis/20171006/"
csv_single_cells <- "Tracked_detectedNeutros.csv"

population <- read_csv(file.path(analysis_folder, csv_single_cells)) %>%
  rename(Metadata_timePoint = Metadata_frame) %>%
  mutate(Metadata_timePoint = as.double(Metadata_timePoint)) 
```

```{r}
single_cell_data <- population %>%
    filter(
      Metadata_condition == "vehi",
      Metadata_id == "bd0037",
      )


```

best track 
pred	20130909	None	gn0048	HEM


2D plot
```{r}
ggplot(
  data = single_cell_data
  ) + 
  aes( 
    x = Location_Center_Y, 
    y = Location_Center_X, 
    colour = AreaShape_Compactness,  
    group = TrackObjects_Label) +
  geom_path() +
  #geom_point() + 
  coord_fixed() +
  theme_void() +
  theme(panel.border = element_rect(size = 1, colour = "grey30", fill = NA ) ) 

  
```

3D plot
```{r}

df <- single_cell_data %>% 
    select(Location_Center_Y, Location_Center_X,TrackObjects_Label, Metadata_timePoint ) %>%
  rename(Zeit = Metadata_timePoint ) %>%
  rename(Standpunkt_X = Location_Center_X) %>% 
  rename(Standpunkt_Y = Location_Center_Y) %>%
  mutate(Zeit = Zeit / 2)

df %>% print(Standpunkt_Y) %>% print
```

```{r}

t <- list(
  family = "sans serif",
  size = 16,
  color = 'blue')

plot_ly(df, 
  x = ~Standpunkt_X, 
  y = ~Standpunkt_Y, 
  z = ~Zeit,  
  split = ~TrackObjects_Label, 
  type = 'scatter3d', 
  mode = 'lines', 
  opacity = 1,
  line = list(width = 2, reverscale = FALSE)
  ) %>% 
    layout(font=t)

```


```{r}
tracks <- single_cell_data %>%
  group_by(TrackObjects_Label) %>%
  track(., c("TrackObjects_Label"))

tracks %>% assess(., min_path_length = 20)
```
