==> 02_track.Rmd <==
---
title: "Prepae metadata file for asthma data"
author: "TB"
date: "7/28/2017"
output: html_document
---
```{r}
library("tidyverse")
library("stringr")
library("readr")
library("plotly")
library("ggplot2")
library("neutrominer")
source("utilities.r")
```

# Load data
The analysis folder contains the segmentation and tracking results. Each 
time lapse experiment is stored in separate folder.
```{r}
analysis_folder <- "../../analysis/sample/20171001_cp3/"
csv_single_cells <- "Tracked_detectedNeutros.csv"
```

# Output 
Results of neutrominer::track are stored in csv_tracks
```{r}
csv_tracks <- 'tracks_cp.csv'
```


# Metadata 
```{r}
strata <- c('Metadata_attractant','Metadata_concentration','Metadata_condition',
            'Metadata_date','Metadata_dose','Metadata_id')
```


# Parameter
To recalculate the speed of the neutrophils in µm/minute we need the pixel size
and the number of frames per minute.
```{r}
# scale speed from pixel/frame to µm/minute
pixel_size = 0.7819
frames_per_second = 2/60
# frames are recorded each 30 seconds, i.e. 
speed_scale <- pixel_size * frames_per_second

track_label = c('TrackObjects_Label')
```


# Load data 
The output of CellProfiler track module is loaded as CSV, tracked using cytominer::track and 
stored in the data folder as "track.csv". 
```{r}
# load csv files from all sub folders 
df <- read_single_cell_data(analysis_folder, csv_single_cells) 

df %>% 
  mutate(Metadata_timePoint = Metadata_frame) %>%
  dplyr::mutate(Metadata_timePoint = as.double(Metadata_timePoint)) 
```
```{r}
df %>% filter(Metadata_id == "kira462") %>%
  ungroup() %>% 
  group_by(Metadata_concentration) %>% 
  slice(1)
```


# remove bad frames
```{r}
clean <- df %>%
  filter( !((Metadata_id == "kira453") & (Metadata_date == "201612a") &  (Metadata_frame < 10) ) )  %>%
  filter( !((Metadata_id == "kira459") & (Metadata_date == "201612b") &  (Metadata_frame < 10) ) ) %>%
  filter( !((Metadata_id == "kira462") & (Metadata_concentration == "1ul") &  (Metadata_frame < 30) ) ) %>%
  filter( !((Metadata_id == "kira456") & (Metadata_date == "201612b") &  (Metadata_frame < 10) ) ) %>%
  filter( !((Metadata_id == "kira456") & (Metadata_date == "201612c") &  (Metadata_frame < 10) ) ) %>%
  print

```

# track the neutros
```{r}
tracks <- clean %>%
  group_by_(.dots =  c(strata,track_label)) %>%
  rename(Metadata_timePoint = Metadata_frame) %>%
  arrange(Metadata_timePoint) %>%
  neutrominer::track(., c(strata,track_label))
```


## PostProcessing 
Calculate speed in µm/minute
```{r}
postprocessed <- tracks %>%
  mutate(Track_Speed = Track_Speed * speed_scale) %>%
  mutate(Track_Speed_max = Track_Speed_max * speed_scale) %>%
  mutate(Track_Speed_X = Track_Speed_X * speed_scale) %>%
  mutate(Track_Speed_Y = Track_Speed_Y * speed_scale) %>%
  mutate(Track_Distance_Traveled = pixel_size * Track_Distance_Traveled) %>%
  mutate(Track_Integrated_Distance_Traveled = pixel_size * Track_Integrated_Distance_Traveled) 
```


# Save data 
* Correct Metadata: all na values are set to 'None'
* store data 
```{r}
postprocessed$Metadata_dose[is.na(postprocessed$Metadata_dose)] <- 'None'
write_csv(postprocessed, file.path(analysis_folder, csv_tracks))
```

```{r}
 quality <- postprocessed %>% 
  group_by_(.dots = strata) %>%
    neutrominer::assess(.,min_path_length = 19) %>%
  arrange(VOT) %>%
  print
```
Two data sets have a low VOT
* kira 099, asthma, ltb4, 201612a
* kira 099, asthma, ltb4, 201612b
* mw, fmlp, 10um, 20170117a
Both time series are characterized by a different image contrast that can not
be processes using the CP pipeline. Both data sets should be removed from further analysis! 

```{r}
quality %>% 
  filter(  !((Metadata_date == "201612a") & (Metadata_id == "kira099") ) ) %>%
  filter(  !((Metadata_date == "201612b") & (Metadata_id == "kira099") ) ) %>%
  filter(  !((Metadata_date == "20170117a") & (Metadata_id == "mw") & (Metadata_attractant == "fmlp") & (Metadata_concentration == "10um") ) ) %>%
  ungroup() %>%
  #group_by(Metadata_condition, Metadata_attractant) %>%
  summarise(mean_vot = mean(VOT), sd_vot = sd(VOT), n = n()) %>%
  print
```

```{r}
quality %>% 
  filter(Metadata_condition == "control-asthma") %>% 
  print
```


```{r}
ggplot(data = quality) +
  aes(x = VOT, y = Exp_Tracks, color = Metadata_condition, shape = Metadata_attractant) +
  geom_point()
```
