==> 05_classify_asthma_morphology.Rmd <==
---
title: "Prepae metadata file for asthma data"
author: "TB"
date: "7/28/2017"
output: html_document
---
```{r}
library("tidyverse")
library("stringr")
library("readr")
library("plotly")
library("ggplot2")
library("neutrominer")
library("magrittr")
library("caret")
library("AppliedPredictiveModeling")
```

# Load data
Analysis folder and tracked data 
```{r}
analysis_folder <- "../../analysis/20171001_cp3/"
csv_profile <- 'profiles.csv'

population <- read_csv(file.path(analysis_folder, csv_profile)) %>%
  filter(Metadata_condition == "asthma" | Metadata_condition == "control-asthma") %>%
  filter(Metadata_attractant == "ltb4") %>%
  filter( !((Metadata_id == "kira453") & (Metadata_date == "201612b")) ) %>%
  filter( !((Metadata_id == "kira095") & (Metadata_date == "201612b")) ) %>%
  filter( !((Metadata_id == "kira094") & (Metadata_date == "201612b"))) %>%
  filter( !((Metadata_date == "201612a") & (Metadata_id == "kira099") ) ) %>%
  filter( !((Metadata_date == "201612b") & (Metadata_id == "kira099") ) ) %>%
  filter( !((Metadata_date == "201612") & (Metadata_id == "kira462") & (Metadata_concentration == "1ul") ) ) 


```
# Metadata 
```{r}
strata <- c('Metadata_attractant','Metadata_concentration','Metadata_condition',
            'Metadata_date','Metadata_dose','Metadata_id')
```

# remove double entries (introduced by CP)
```{r}

feature_cols <- base::setdiff(
  colnames(population) %>%
    stringr::str_subset("^Area|^Radial|^Texture"),
  colnames(population) %>%
    stringr::str_subset("_1")
)

feature_cols <- setdiff(feature_cols, "TrackObjects_Label")
```

# remove na columns
```{r}
na_colnames <- cytominer::drop_na_columns(
  population = population,
  variables = feature_cols,
  cutoff = 0)

feature_cols <- setdiff(feature_cols, na_colnames)

clean <- cytominer::drop_na_rows(population = population, variables = feature_cols)
```

# remove zero variance features
```{r}

null_variance_features <- cytominer::variance_threshold(
  sample = clean, 
  variables = feature_cols
  ) %>%
  print

feature_cols <- setdiff(feature_cols, null_variance_features)

used_feature_names <- setdiff(feature_cols,
clean %>% 
  ungroup() %>% 
  select_(.dots = feature_cols) %>%
  as.matrix() %>%
  cor(.) %>%
  findCorrelation(.,cutoff =0.95, names = TRUE)
)



```

Aggregate data for each experiment
```{r}
aggregated <-
  cytominer::aggregate(
    population = clean,
    variables = used_feature_names,
    strata = c('Metadata_attractant','Metadata_concentration','Metadata_condition','Metadata_dose','Metadata_id')
  )
```

```{r}
asthma <- aggregated %>% 
  ungroup() %>%
  dplyr::select_(.dots = c("Metadata_condition", used_feature_names )) %>%
  dplyr::mutate(Metadata_condition = as.factor(Metadata_condition)) %>%
  dplyr::rename(class = Metadata_condition) %>%
   dplyr::select(
   -RadialDistribution_ZernikeMagnitude_neutrosIlluCorrected_6_6,
  # -RadialDistribution_ZernikeMagnitude_neutrosIlluCorrected_7_7,
   -RadialDistribution_ZernikeMagnitude_neutrosIlluCorrected_8_8,
   -RadialDistribution_ZernikeMagnitude_neutrosIlluCorrected_9_9,
  # -RadialDistribution_ZernikeMagnitude_neutrosIlluCorrected_5_5
  #RadialDistribution_ZernikeMagnitude_neutrosIlluCorrected_9_9
    ) %>%
  print
 
```
Create simple test data set 
```{r}

cnames <- colnames(asthma)

set.seed(123)

train_control = trainControl(method = "LOOCV")
  
lda_model <- train(class~., data = asthma, 
                 method = "lda", 
                 trControl = train_control,
                 verbose = FALSE)
lda_model
```


```{r}
data_asthma <- clean %>%
  ungroup() %>%
  select_(.dots = c('Metadata_condition',used_feature_names)) %>%
  mutate(Metadata_condition = as.factor(Metadata_condition))

res.man <- manova(as.matrix(data_asthma[,used_feature_names]) ~ Metadata_condition, data = data_asthma) %>% print

summary(res.man)
```
